+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Student tabs feature, saving the lesson state
 - Improved units diagram usability: added tooltips, responsive diagram size and clicks prevention while loading
 - Student answers also depends on course and block
+- Icon shows when student has answered a question
 
 ### Changed