+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When editing cards its not allowed to drag elements
 - When editing card tittle, the item wont toggle
 - Deleting lesson doesnt return error.
+- Questions do not disorder in student lessons
 
 ## [1.0.1]