+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Student tabs feature, saving the lesson state
 - Improved units diagram usability: added tooltips, responsive diagram size and clicks prevention while loading
 - Student answers also depends on course and block
+- Icon shows when student has answered a question
 
 ### Changed
 
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When editing cards its not allowed to drag elements
 - When editing card tittle, the item wont toggle
 - Deleting lesson doesnt return error.
+- Questions do not disorder in student lessons
 
 ## [1.0.1]